Backend/requirements.txt
CacheControl
chardet
colorama
contextlib2
distlib
distro
docopt
html5lib
idna
lxml
mechanize
msgpack
npyscreen
ordered-set
packaging
paho-mqtt
pep517
ply
progress
pycparser
pyparsing
pytz
PyYAML
toml
udiskie
webencodings
urllib3
requests
idna
snscrape

